archive/requirements.txt
requests==2.27.1
environs==9.5.0
cryptography==37.0.1
pyjwt==2.3.0
pyOpenSSL==22.0.0
pytest==7.1.2